Write to database functionality
We will be creating a new branch and new PR for promisesChecker. We have addressed the changes we could.
