--- src/assets/LOGIN_wCHAR_666.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d926ba9bfdccc22cebf3f0b0fce3e1974c800d7a80b6ba969101e86377a7b92f
-size 57837
