+++ src/components/ExperimentalTask.js
@@ -3,6 +3,7 @@ import { Graphics, Text, useApp } from "@inlet/react-pixi";
 import CursorMode from "./CursorMode.js";
 import Pose from "./Pose/index";
 import { white, darkGray, yellow } from "../utils/colors";
+import { writeToDatabase, promiseChecker } from "../firebase/database.js";
 
 const ExperimentalTask = (props) => {
   const {
@@ -12,7 +13,7 @@ const ExperimentalTask = (props) => {
     onComplete,
     rowDimensions,
     cursorTimer,
-    currentConjectureIdx
+    currentConjectureIdx,
   } = props;
   const [showCursor, setShowCursor] = useState(false);
 
@@ -25,7 +26,41 @@ const ExperimentalTask = (props) => {
     g.drawRect(col3.x, col3.y, col3.width, col3.height);
     g.endFill();
   });
-  
+
+  // Database Write Functionality
+  // This lives in ExperimentalTask.js so it only runs when the user is participating in a conjecture
+  useEffect(() => {
+    // Optional URL parameters for whether motion data recording is enabled
+    // and what the fps is for recording.
+    // Defaults are false and 30.
+    const queryParameters = new URLSearchParams(window.location.search);
+    const recordingUrlParam = queryParameters.get("recording") || "false";
+    // If the recording param is set to true, run the recording functions
+    if (recordingUrlParam.toLowerCase() === "true") {
+      // Get the query parameter. If empty, set to 30 for a default.
+      const fpsUrlParam = parseInt(queryParameters.get("fps")) || 30;
+      // Empty array to hold the promise objects.
+      // This is important so we can assure that all the promises get settled on component unmount.
+      let promises = [];
+      // This creates an interval for the writing to the database every n times a second, where n is a variable framerate.
+      const intervalId = setInterval(() => {
+        // Call the writeToDatabase function
+        // Push the resulting promise object to the promises array
+        promises.push(
+          writeToDatabase(poseData, currentConjectureIdx, fpsUrlParam)
+        );
+      }, 1000 / fpsUrlParam);
+
+      return async () => {
+        // On component unmount:
+        // Stop the interval
+        clearInterval(intervalId);
+        // Wait until all promises are settled so we don't lose data
+        await Promise.allSettled(promises);
+      };
+    }
+  }, []);
+
   useEffect(() => {
     const timeout = setTimeout(
       () => {
